HS-248/back/feature/word game/result/get 사용자는 낱말 게임이 끝나면 게임 참여자들의 점수와 등수를 최종적으로 볼 수 있다.

See merge request s07-webmobile1-sub2/S07P12D109!88

+++ backend/src/main/java/com/hans/hans/domain/wordgame/dto/WordGameAnswerResponseDto.java
@@ -9,12 +9,12 @@ import java.util.Map;
 public class WordGameAnswerResponseDto {
 
     private Long roomSequence;
-    private Map<String, Integer> correctPlayers;
+    private Map<String, Long> correctPlayers;
     private int point;
     private String answer;
 
     @Builder
-    public WordGameAnswerResponseDto(Long roomSequence, Map<String, Integer> correctPlayers, int point, String answer){
+    public WordGameAnswerResponseDto(Long roomSequence, Map<String, Long> correctPlayers, int point, String answer){
         this.roomSequence = roomSequence;
         this.correctPlayers = correctPlayers;
         this.point = point;
